pyproject.toml
[tool.poetry]
name = "chatgpt-clone"
version = "0.1.0"
description = ""
authors = ["kamran945 <kamran945@gmail.com>"]
readme = "README.md"
packages = [
    { include = "src" },
]

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
langchain = "^0.3.7"
langchain-huggingface = "^0.1.2"
langchain-groq = "^0.2.1"
streamlit = "^1.40.0"
langchain-core = "^0.3.15"
python-dotenv = "^1.0.1"
langchain-ollama = "^0.2.0"
langgraph = "^0.2.53"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
